Use prototype inheritance for Task components
